requirements.txt
# Core CV
opencv-python>=4.8.0
ultralytics>=8.0.0

# Deep SORT tracking
deep-sort-realtime>=1.3.2

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Dashboard
streamlit>=1.28.0
plotly>=5.18.0

# Data
pandas>=2.0.0
numpy>=1.24.0

# Database
SQLAlchemy>=2.0.0

# Utilities
python-multipart>=0.0.6
requests>=2.31.0
loguru>=0.7.0
PyYAML>=6.0.1
tqdm>=4.66.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
